+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - When validating boot sets, if a boot artifact is missing from a manifest, include the manifest S3 URL
   in the associated message.
 - Add check that at least one hardware-specifier field is non-empty to `validate_sanitize_boot_sets`
+- When validating boot sets (at session template patch/creation/validation or session creation),
+  attempt to validate that the specified architecture matches that of the corresponding IMS image.
+  Mismatches are fatal errors.
 
 ### Changed
 - Marked PATCH session status endpoint to be ignored by the CLI.